requirements.txt
alembic==1.15.2
annotated-types==0.7.0
anyio==4.9.0
click==8.1.8
exceptiongroup==1.2.2
fastapi==0.115.12
filelock==3.18.0
fsspec==2025.3.2
h11==0.14.0
idna==3.10
iniconfig==2.1.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.2.1
numpy==2.0.2
packaging==25.0
pandas==2.2.3
pluggy==1.5.0
pydantic==2.11.3
pydantic_core==2.33.1
pytest==8.3.5
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytz==2025.2
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.46.2
sympy==1.13.3
tomli==2.2.1
torch==2.7.0
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
uvicorn==0.34.2
